add poetry package manager and docker-compose. Fix environments in settings file
