+++ src/samplesite/wsgi.py
@@ -1,0 +1,16 @@
+"""
+WSGI config for samplesite project.
+
+It exposes the WSGI callable as a module-level variable named ``application``.
+
+For more information on this file, see
+https://docs.djangoproject.com/en/5.1/howto/deployment/wsgi/
+"""
+
+import os
+
+from django.core.wsgi import get_wsgi_application
+
+os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'samplesite.settings')
+
+application = get_wsgi_application()
